+++ ILEditor/ILEditor.csproj
@@ -133,7 +133,10 @@
     <Compile Include="Classes\IBMi.cs" />
     <Compile Include="Classes\IBMiUtils.cs" />
     <Compile Include="Classes\LanguageTools\CLFile.cs" />
+    <Compile Include="Classes\LanguageTools\RPGLexer.cs" />
+    <Compile Include="Classes\LanguageTools\RPGFree.cs" />
     <Compile Include="Classes\LanguageTools\RPGParser.cs" />
+    <Compile Include="Classes\LanguageTools\Types.cs" />
     <Compile Include="Classes\RemoteSource.cs" />
     <Compile Include="Classes\ILEObject.cs" />
     <Compile Include="Classes\MemberCache.cs" />
@@ -315,6 +318,12 @@
     <Compile Include="UserTools\ObjectBrowse.Designer.cs">
       <DependentUpon>ObjectBrowse.cs</DependentUpon>
     </Compile>
+    <Compile Include="UserTools\OutlineView.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="UserTools\OutlineView.Designer.cs">
+      <DependentUpon>OutlineView.cs</DependentUpon>
+    </Compile>
     <Compile Include="UserTools\SourceEditor.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -450,6 +459,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="UserTools\ObjectBrowse.resx">
       <DependentUpon>ObjectBrowse.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="UserTools\OutlineView.resx">
+      <DependentUpon>OutlineView.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="UserTools\SourceEditor.resx">
       <DependentUpon>SourceEditor.cs</DependentUpon>
